+++ webapp/tests/unit/controller/handlebar-appconfig-alwaysallow.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for W-129: appConfig.system allowlist exposure (unauthenticated)
  * @description     Tests for contextFilter.alwaysAllow to re-expose specific appConfig.system keys after filtering
  * @file            webapp/tests/unit/controller/handlebar-appconfig-alwaysallow.test.js
- * @version         1.6.9
- * @release         2026-02-06
+ * @version         1.6.10
+ * @release         2026-02-07
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/