+++ webapp/tests/unit/i18n/find-dynamic-keys.js
@@ -3,8 +3,8 @@
  * @tagline         Helper script to find dynamic i18n key usage
  * @description     Standalone script to search for dynamic i18n key patterns
  * @file            webapp/tests/unit/i18n/find-dynamic-keys.js
- * @version         1.6.9
- * @release         2026-02-06
+ * @version         1.6.10
+ * @release         2026-02-07
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/