+++ webapp/tests/unit/site/hello-vue-controller.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit tests for HelloVue SPA demonstration controller
  * @description     Unit tests for HelloVue SPA demonstration controller
  * @file            webapp/tests/unit/site/hello-vue-controller.test.js
- * @version         1.6.9
- * @release         2026-02-06
+ * @version         1.6.10
+ * @release         2026-02-07
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/