+++ webapp/tests/unit/utils/jpulse-ui-input-taginput.test.js
@@ -3,8 +3,8 @@
  * @tagline         Unit Tests for jPulse.UI.input.tagInput (W-148)
  * @description     Tests for tagInput: parseValue, formatValue, init/sync
  * @file            webapp/tests/unit/utils/jpulse-ui-input-taginput.test.js
- * @version         1.6.9
- * @release         2026-02-06
+ * @version         1.6.10
+ * @release         2026-02-07
  * @repository      https://github.com/jpulse-net/jpulse-framework
  * @author          Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/
  * @copyright       2025-2026 Peter Thoeny, https://twiki.org & https://github.com/peterthoeny/